feat(#1411): CLEAN: Split adapter.ts (137-line handleContextInfo + 691 nbnc file)
